+++ brikar-client/src/main/java/com/truward/brikar/client/rest/RestOperationsFactory.java
@@ -1,12 +1,8 @@
-package com.truward.brikar.client.rest.support;
+package com.truward.brikar.client.rest;
 
-import com.truward.brikar.client.binder.RestServiceBinder;
-import com.truward.brikar.client.binder.RestServiceBinderFactory;
 import com.truward.brikar.client.interceptor.RequestLogAwareHttpRequestInterceptor;
 import com.truward.brikar.client.interceptor.RequestLogAwareHttpResponseInterceptor;
-import com.truward.brikar.client.rest.RestClientBuilderFactory;
-import com.truward.brikar.client.rest.RestClientBuilder;
-import org.apache.http.*;
+import org.apache.http.HttpHost;
 import org.apache.http.auth.AuthScope;
 import org.apache.http.auth.UsernamePasswordCredentials;
 import org.apache.http.client.CredentialsProvider;
@@ -16,11 +12,10 @@ import org.apache.http.impl.client.DefaultHttpRequestRetryHandler;
 import org.apache.http.impl.client.HttpClientBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.BeanInitializationException;
 import org.springframework.beans.factory.DisposableBean;
-import org.springframework.beans.factory.InitializingBean;
 import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
 import org.springframework.http.converter.HttpMessageConverter;
+import org.springframework.web.client.RestOperations;
 import org.springframework.web.client.RestTemplate;
 
 import javax.annotation.Nonnull;
@@ -31,11 +26,12 @@ import java.util.List;
 import java.util.concurrent.TimeUnit;
 
 /**
- * Default implementation of {@link RestClientBuilderFactory}, adapted for spring framework.
+ * A helper class that pre configures RestOperations and underlying HTTP client to use the specified connection
+ * settings.
  *
  * @author Alexander Shabanov
  */
-public class StandardRestClientBuilderFactory implements RestClientBuilderFactory, InitializingBean, DisposableBean, AutoCloseable {
+public final class RestOperationsFactory implements DisposableBean, AutoCloseable {
   /**
    * Default connection ttl setting. Should be less than server we're going to use.
    * Since this class is intended to be used mostly for interacting with brikar services, this TTL
@@ -53,30 +49,26 @@ public class StandardRestClientBuilderFactory implements RestClientBuilderFactor
    */
   public static final int DEFAULT_MAX_CONN_TOTAL = 15;
 
-  private final Logger log = LoggerFactory.getLogger(getClass());
+  /**
+   * Use logger with specified name - 'BrikarRestClient' that should be associated with all HTTP calls.
+   */
+  private final Logger log = LoggerFactory.getLogger("BrikarRestClient");
+
   private final CredentialsProvider credentialsProvider = new BasicCredentialsProvider();
-  private RestServiceBinder restServiceBinder;
-  private RestServiceBinderFactory restServiceBinderFactory;
   private final List<HttpMessageConverter<?>> messageConverters;
   private HttpComponentsClientHttpRequestFactory httpRequestFactory;
   private long connectionTtlMillis;
   private int maxConnTotal;
   private HttpRequestRetryHandler retryHandler;
 
-  public StandardRestClientBuilderFactory(@Nonnull List<HttpMessageConverter<?>> messageConverters,
-                                          @Nonnull RestServiceBinderFactory restServiceBinderFactory) {
+  public RestOperationsFactory(@Nonnull List<HttpMessageConverter<?>> messageConverters) {
     this.messageConverters = messageConverters;
-    setRestServiceBinderFactory(restServiceBinderFactory);
     setConnectionTtlMillis(DEFAULT_CONNECTION_TTL);
     setMaxConnTotal(DEFAULT_MAX_CONN_TOTAL);
     setRetryHandler(null);
   }
 
-  public StandardRestClientBuilderFactory(@Nonnull List<HttpMessageConverter<?>> messageConverters) {
-    this(messageConverters, RestServiceBinderFactory.DEFAULT);
-  }
-
-  public StandardRestClientBuilderFactory(@Nonnull HttpMessageConverter<?>... messageConverters) {
+  public RestOperationsFactory(@Nonnull HttpMessageConverter<?>... messageConverters) {
     this(Arrays.asList(messageConverters));
   }
 
@@ -85,7 +77,7 @@ public class StandardRestClientBuilderFactory implements RestClientBuilderFactor
    * that should be given in milliseconds.
    * If the provided value is negative, then no value will be set and default http client connection TTL settings will
    * be used which sets TTL to infitity.
-   * <p>Should be set prior to {@link #afterPropertiesSet()}.</p>
+   * <p>Should be set prior to {@link #getRestOperations()} to take an effect.</p>
    *
    * @param connectionTtlMillis Connection time to live, in milliseconds
    */
@@ -95,7 +87,7 @@ public class StandardRestClientBuilderFactory implements RestClientBuilderFactor
 
   /**
    * If set before bean is initialized, instructs to use provided value as maximum connections.
-   * <p>Should be set prior to {@link #afterPropertiesSet()}.</p>
+   * <p>Should be set prior to {@link #getRestOperations()} to take an effect.</p>
    *
    * @param maxConnTotal Maximum total connections
    */
@@ -105,7 +97,7 @@ public class StandardRestClientBuilderFactory implements RestClientBuilderFactor
 
   /**
    * Sets retry handler, that should be used for handling connectivity issues.
-   * <p>Should be set prior to {@link #afterPropertiesSet()}.</p>
+   * <p>Should be set prior to {@link #getRestOperations()} to take an effect.</p>
    *
    * @param retryHandler Retry handler
    */
@@ -114,35 +106,36 @@ public class StandardRestClientBuilderFactory implements RestClientBuilderFactor
   }
 
   /**
-   * Sets service binder, that will create a service by given class and RestOperations instance.
-   * <p>Should be set prior to {@link #afterPropertiesSet()}.</p>
-   * <p>See also {@link RestServiceBinder}</p>
+   * Sets credentials for use when accessing service APIs.
+   * Note, that multiple calls to this method override previously set credentials.
    *
-   * @param factory Factory instance
+   * @param credentials Credential list to use
    */
-  public void setRestServiceBinderFactory(@Nonnull RestServiceBinderFactory factory) {
-    this.restServiceBinderFactory = factory;
-  }
+  public void setCredentials(@Nonnull List<ServiceClientCredentials> credentials) {
+    credentialsProvider.clear();
 
-  @Nonnull
-  @Override
-  public <T> RestClientBuilder<T> newClient(@Nonnull Class<T> serviceClass) {
-    if (restServiceBinder == null) {
-      throw new IllegalStateException("Bean has not been initialized properly");
+    for (final ServiceClientCredentials cred : credentials) {
+      final URI uri = cred.getBaseUri();
+      credentialsProvider.setCredentials(new AuthScope(
+              new HttpHost(uri.getHost(), uri.getPort(), uri.getScheme())),
+          new UsernamePasswordCredentials(cred.getUsername(), cred.getPassword().toString()));
     }
-    return new InternalRestClientBuilder<>(serviceClass, restServiceBinder, credentialsProvider);
   }
 
-  @Override
-  public void afterPropertiesSet() {
+  /**
+   * Returns rest operations object, suitable for use with the specified credentials
+   *
+   * @return New instance of {@link RestOperations} with all the connection settings provided prior to calling this
+   */
+  @Nonnull
+  public RestOperations getRestOperations() {
     final HttpClientBuilder builder = HttpClientBuilder.create();
     initDefaultHttpClientBuilder(builder);
 
     httpRequestFactory = new HttpComponentsClientHttpRequestFactory(builder.build());
     final RestTemplate restTemplate = new RestTemplate(httpRequestFactory);
     initRestTemplate(restTemplate);
-
-    restServiceBinder = restServiceBinderFactory.create(restTemplate);
+    return restTemplate;
   }
 
   @Override
@@ -204,68 +197,5 @@ public class StandardRestClientBuilderFactory implements RestClientBuilderFactor
 
   protected void initRestTemplate(@Nonnull RestTemplate restTemplate) {
     restTemplate.setMessageConverters(messageConverters);
-  }
-
-  //
-  // Private
-  //
-
-  private static final class InternalRestClientBuilder<T> implements RestClientBuilder<T> {
-    private String username;
-    private String password;
-    private URI restServiceUri;
-    private final RestServiceBinder restServiceBinder;
-    private final CredentialsProvider credentialsProvider;
-    private final Class<T> clazz;
-
-    public InternalRestClientBuilder(@Nonnull Class<T> clazz,
-                                     @Nonnull RestServiceBinder restServiceBinder,
-                                     @Nonnull CredentialsProvider credentialsProvider) {
-      this.clazz = clazz;
-      this.restServiceBinder = restServiceBinder;
-      this.credentialsProvider = credentialsProvider;
-    }
-
-    @Nonnull
-    @Override
-    public RestClientBuilder<T> setUri(@Nonnull URI uri) {
-      this.restServiceUri = uri;
-      return this;
-    }
-
-    @Nonnull
-    @Override
-    public RestClientBuilder<T> setUsername(@Nullable String username) {
-      this.username = username;
-      return this;
-    }
-
-    @Nonnull
-    @Override
-    public RestClientBuilder<T> setPassword(@Nullable String password) {
-      this.password = password;
-      return this;
-    }
-
-    @Nonnull
-    @Override
-    public T build() {
-      final String username = this.username;
-      final String password = this.password;
-      final URI uri = this.restServiceUri;
-
-      if (uri == null) {
-        throw new BeanInitializationException("uri has not been set");
-      }
-
-      if (username != null && password != null) {
-        credentialsProvider.setCredentials(new AuthScope(
-                new HttpHost(uri.getHost(), uri.getPort(), uri.getScheme())),
-            new UsernamePasswordCredentials(username, password));
-      }
-
-      final String url = restServiceUri.toString();
-      return restServiceBinder.createClient(url, clazz);
-    }
   }
 }