+++ brikar-client/src/main/java/com/truward/brikar/client/rest/RestServiceBinder.java
@@ -1,6 +1,7 @@
-package com.truward.brikar.client.binder;
+package com.truward.brikar.client.rest;
 
 import javax.annotation.Nonnull;
+import java.net.URI;
 
 /**
  * @author Alexander Shabanov
@@ -9,6 +10,11 @@ public interface RestServiceBinder {
 
   @Nonnull
   <T> T createClient(@Nonnull String serviceBaseUrl,
+                     @Nonnull Class<T> restServiceClass,
+                     @Nonnull Class<?> ... extraClasses);
+
+  @Nonnull
+  <T> T createClient(@Nonnull URI serviceBaseUrl,
                      @Nonnull Class<T> restServiceClass,
                      @Nonnull Class<?> ... extraClasses);
 }